Changed the framework of the bot sompletly
the past frame work was giving a lot of trouble so I decided to completely abandon it and change everything
