--- config.json
@@ -1,17 +1,0 @@
-{
-    "intervals": {
-        "min": 10,
-        "max": 25
-    },
-    "channel_ids" :[
-        "",
-        "",
-        "",
-        "",
-        ""
-    ],
-    "token" :"",
-    "token_1" :""
-
-
-}
