--- example.json
@@ -1,20 +1,0 @@
-{
-    //Enter ywo intervals here,they will be taken in sec in default
-    "intervals": {
-        "min": 10,
-        "max": 25
-    },
-    //Adding as many channals as you can and leave the rest
-    "channel_ids" :[
-        "YOUR-CHANNEL-IDS_fill_as_many_as_you_can",
-        "THE_MORE_THE_BETTER",
-        "",
-        "",
-        ""
-    ],
-    //paste your token here
-    "token" :"PASTE-YOUR-TOKEN-HERE",
-    "token_1" :"Your alt token"
-
-
-}
