+++ install.bat
@@ -1,0 +1,4 @@
+@echo off
+cd /d %~dp0
+pip install -r requirements.txt
+pause
