+++ run.bat
@@ -1,0 +1,5 @@
+@echo off
+cd /d %~dp0
+python bot.py
+python extra.py
+pause
