--- starter.bat
@@ -1,3 +1,0 @@
-@echo off
-start "" main.py
-start "" extra.py
